added data checker button
Add a date check button into Abdul GSX added a new file called data check a dialogue.GSX
This now works you can click the data check button in the top right hand corner and it brings up all of the information that the model is using in order to produce the advice. This lets the user ensure that it's using accurate information in order to compete the advice at the en.